# docs/swagger.yaml
basePath: /
definitions:
...
  model.Organization:
    properties:
      createdAt:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.RequestMemberUpdateOther:
    properties:
      role:
        type: string
    type: object
  model.RequestMemberUpdateSelf:
    properties:
      name:
        type: string
      newPassword:
        maxLength: 100
        minLength: 8
        type: string
      oldPassword:
        maxLength: 100
        minLength: 8
        type: string
    type: object
  model.RequestOrganizationCreate:
    properties:
      email:
        type: string
      name:
        type: string
      organizationName:
        type: string
    required:
    - email
    - name
    - organizationName
    type: object
  model.RequestOrganizationUpdate:
    properties:
      description:
        type: string
      organizationName:
        type: string
    type: object
  model.RequestPasswordReset:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
      resetToken:
        type: string
    required:
    - password
    - resetToken
    type: object
  model.RequestPasswordResetEmail:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  model.RequestTransferOwnership:
    properties:
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - password
    type: object
...
  repository.OrganizationMemberWithRole:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
host: platform-api.string-api.xyz
info:
  contact:
    email: support@stringxyz.com
    name: Platform Management API Support
    url: http://string.xyz
  description: Platform Management API for managing Organizations, Platforms, API
    keys, etc.
  title: Platform Management API
  version: "1.0"
paths:
...
  /members:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.OrganizationMemberWithRole'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get all members
      tags:
      - Member
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update Self Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestMemberUpdateSelf'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update self
      tags:
      - Member
  /members/{id}:
    get:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get a member by ID
      tags:
      - Member
    patch:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Member Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestMemberUpdateOther'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update member
      tags:
      - Member
  /members/{id}/deactivate:
    patch:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Deactivate member
      tags:
      - Member
  /members/{id}/reactivate:
    patch:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Reactivate member
      tags:
      - Member
  /members/{id}/transferOwner:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: string
      - description: Transfer Ownership Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestTransferOwnership'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.OrganizationMemberWithRole'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Transfer ownership
      tags:
      - Member
  /members/password-reset:
    get:
      consumes:
      - application/json
      parameters:
      - description: Password Reset Email Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestPasswordResetEmail'
      produces:
      - application/json
      responses:
        "200":
          description: Email sent message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Send password reset email
      tags:
      - Member
    post:
      consumes:
      - application/json
      parameters:
      - description: Password Reset Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestPasswordReset'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Reset password
      tags:
      - Member
...
  /organizations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Organization'
        "500":
          description: Internal Server Error
          schema:
            type: integer
      summary: Get organization
      tags:
      - Organization
    patch:
      consumes:
      - application/json
      parameters:
      - description: Organization Update Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestOrganizationUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Organization'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update organization
      tags:
      - Organization
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization Create Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RequestOrganizationCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Organization'
        "400":
          description: Bad Request
          schema:
            type: integer
        "409":
          description: Conflict
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      summary: Create organization
      tags:
      - Organization
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: StringAdminJWT
  RefreshBearerAuth:
    type: apiKey
    in: header
    name: StringAdminRefreshToken
  JWTAuth:
    type: apiKey
    in: cookie
    name: StringAdminJWT
  RefreshTokenAuth:
    type: apiKey
    in: cookie
    name: StringAdminRefreshToken
swagger: "2.0"